Chgagnon/fix load issue (#11422)
* Remove specific implementation references

* Move SlickTableDataView into own module

* Add new module
