+++ src/sql/base/browser/ui/table/tableDataView.ts
@@ -254,34 +254,3 @@ export class TableDataView<T extends Slick.SlickData> implements IDisposableData
 		this._findArray = [];
 	}
 }
-
-export class SlickTableDataView<T extends Slick.SlickData> extends Slick.Data.DataView<T>
-	implements IDisposableDataProvider<T>
-{
-	push(items: Array<T>): void;
-	push(item: T): void;
-	push(input: T | Array<T>): void {
-		let inputArray = new Array();
-		if (Array.isArray(input)) {
-			inputArray.push(...input);
-		} else {
-			inputArray.push(input);
-		}
-
-		this.beginUpdate();
-		inputArray.forEach(item => this.addItem(item));
-		this.endUpdate();
-	}
-
-	public setData(data: T[]): void {
-		let inputArray = new Array();
-		inputArray.push(...data);
-		this.setItems(inputArray);
-	}
-
-	dispose(): void {
-
-	}
-
-}
-
